Fixes 3831 and 3839: Local vsix file and Artifacts with symbol files without versions (#3868)
This is a reduced version of the former closed PR #3834.

- Allow vsixFile to be a local file for New-BCCompilerFolder (#3831)
- Fixes issues determining test symbols for W1 and OnPrem artifacts not
containing versions in their name (#3839)

After that, there is still an issue with RunAlTool always downloading
the AL Language extension. There should be a check or an additional
parameter as hint for an existing vsix or extracted vsix.

Regarding the other PR: I would recommend modifying the caching paths
(in AL-Go) a bit to allow caching of the single vsix in a different base
path to not mixing up caching. When specified, this is most likely a
different version compared to the rest of the compiler folder contents.

+++ CompilerFolderHandling/New-BcCompilerFolder.ps1
@@ -199,13 +199,36 @@ try {
     if ($vsixFile) {
         # If a vsix file was specified unpack directly to compilerfolder
         Write-Host "Using $vsixFile"
-        $tempZip = Join-Path ([System.IO.Path]::GetTempPath()) "alc.$containerName.zip"
-        Download-File -sourceUrl $vsixFile -destinationFile $tempZip
-        Expand-7zipArchive -Path $tempZip -DestinationPath $containerCompilerPath
+
+        if ($vsixFile -notlike 'https://*') {
+            if (Test-Path -Path $vsixFile -PathType Leaf) {
+                Expand-7zipArchive -Path $vsixFile -DestinationPath $containerCompilerPath
+            }
+
+            elseif (Test-Path -Path $vsixFile -PathType Container) {
+                if (!(Test-Path -Path $containerCompilerPath)) {
+                    New-Item $containerCompilerPath -ItemType Directory | Out-Null
+                }
+
+                Copy-Item -Path (Join-Path -Path $vsixFile -ChildPath '*') -Destination $containerCompilerPath -Recurse -Force
+            }
+
+            else {
+                throw "An invalid vsix file was specified."
+            }
+        }
+        else {
+            $tempZip = Join-Path ([System.IO.Path]::GetTempPath()) "alc.$containerName.zip"
+
+            Download-File -sourceUrl $vsixFile -destinationFile $tempZip
+            Expand-7zipArchive -Path $tempZip -DestinationPath $containerCompilerPath
+
+            Remove-Item -Path $tempZip -Force -ErrorAction SilentlyContinue
+        }
+
         if ($isWindows -and $newtonSoftDllPath) {
             Copy-Item -Path $newtonSoftDllPath -Destination (Join-Path $containerCompilerPath 'extension\bin') -Force -ErrorAction SilentlyContinue
         }
-        Remove-Item -Path $tempZip -Force -ErrorAction SilentlyContinue
     }
 
     # If a cacheFolder was specified, the cache folder has been populated
